requirements.txt
absl-py==0.2.2
aniso8601==3.0.0
asn1crypto==0.24.0
astor==0.6.2
bcrypt==3.1.4
bleach==1.5.0
certifi==2018.4.16
cffi==1.11.5
click==6.7
cryptography==2.3
fabric==2.2.1
Flask==1.0.2
flask-restplus==0.10.1
gast==0.2.0
grpcio==1.13.0
gunicorn==19.9.0
html5lib==0.9999999
idna==2.7
invoke==1.1.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.5
paramiko==2.4.1
protobuf==3.6.0
pyasn1==0.4.3
pycparser==2.18
PyNaCl==1.2.1
pytz==2018.4
PyYAML==3.12
six==1.11.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
